Multi-drop move and simulation mode
- Adds support for move_drops method, a replacement for move_drop that
supports up to 5 drops at once using group capacitance for feedback instead
of active.
- Adds --sim option and simplistic drop motion model
- Refactors PurpleDropDevice so that SerialPurpleDropDevice and
SimulatedPurpleDropDevice can share it as a base
- Separates PurpleDropController into controller.py to address circular
dependency, and just because purpledrop.py was getting huge.
- Improves error messaging when trying to send messages via API without
a device connected
- Refactors SyncListener into a ContextManager, because cleaning up in
__del__ method was resulting in intermittent errors, e.g.:
"gevent.exceptions.BlockingSwitchOutError: Impossible to call blocking function in the event loop callback"
