+++ purpledrop/script/pd_server.py
@@ -8,8 +8,10 @@ import os
 
 from purpledrop.calibration import load_electrode_offset_calibration
 from purpledrop.electrode_board import load_board
-from purpledrop.purpledrop import PersistentPurpleDropDevice, PurpleDropController
+from purpledrop.purpledrop import PersistentPurpleDropDevice
+from purpledrop.controller import PurpleDropController
 from purpledrop.playback import PlaybackPurpleDrop, index_log
+from purpledrop.simulated_purpledrop import SimulatedPurpleDropDevice
 from purpledrop.video_client import VideoClientProtobuf
 import purpledrop.server as server
 
@@ -18,11 +20,24 @@ import purpledrop.server as server
 @click.option('--board', 'board_file', help='Board name or path to board definition JSON file', default='misl_v4')
 @click.option('--ecal', 'electrode_calibration_file', help='Name of calibration or path to JSON file', required=False)
 @click.option('--replay', 'replay_file', help='Launch replay server instead of connecting to HW', required=False)
-def main(verbose, board_file, replay_file, electrode_calibration_file=None):
+@click.option('--sim', help='Simulate a purpledrop device', required=False)
+def main(verbose, board_file, replay_file, sim, electrode_calibration_file=None, ):
     """Runs hardware gateway
 
     Will auto-connect to any detected purpledrop USB devices, and provides HTTP interfaces for control.
-    Optionally, supports replay mode in which a recorded event stream is played back
+    Optionally, supports replay and simulation mode.
+
+    In replay mode, a recorded event stream is played back.
+
+    In sim mode, a simple simulated purpledrop device replaces actual hardware
+    connection to allow for testing control scripts. Drops may be populated on
+    some electrodes by providing a list of pins to the --sim argument. The
+    simulator implements a very simple drop model, where each electrode is either
+    empty or covered, and a drop on an inactive electrode can move to an active
+    neighboring electrode. Only grid electrodes are supported (e.g. no reservoirs).
+
+    For example: `pdserver --sim 5,10,120` to create a simulation starting with
+    drops present on electrodes 5, 10, and 120.
     """
     if verbose == 0:
         console_log_level = logging.WARNING
@@ -38,18 +53,24 @@ def main(verbose, board_file, replay_file, electrode_calibration_file=None):
         datefmt="%H:%M:%S",
         level=console_log_level)
 
+    if replay_file is not None and sim is not None:
+        print("Cannot run both --replay and --sim")
+        sys.exit(-1)
+
     board = load_board(board_file)
     if board is None:
         print(f"Could not load board definition for {board_file}")
         sys.exit(1)
 
-    if electrode_calibration_file is not None:
-        print(f"Loading calibration from {electrode_calibration_file}")
-        ecal = load_electrode_offset_calibration(electrode_calibration_file)
-    elif not os.path.exists(board_file):
-        ecal = load_electrode_offset_calibration(board_file)
-        if ecal is not None:
-            print(f"Loading calibration based on {board_file} board name")
+    ecal = None
+    if sim is None:
+        if electrode_calibration_file is not None:
+            print(f"Loading calibration from {electrode_calibration_file}")
+            ecal = load_electrode_offset_calibration(electrode_calibration_file)
+        elif not os.path.exists(board_file):
+            ecal = load_electrode_offset_calibration(board_file)
+            if ecal is not None:
+                print(f"Loading calibration based on {board_file} board name")
 
     video_client = None
     if replay_file is not None:
@@ -59,6 +80,10 @@ def main(verbose, board_file, replay_file, electrode_calibration_file=None):
         print(f"Done. Loaded {end_time - start_time} seconds of data.")
         print("Launching replay server...")
         pd_control = PlaybackPurpleDrop(replay_file, index, board)
+    elif sim:
+        drop_pins = [int(x) for x in sim.split(',')]
+        pd_dev = SimulatedPurpleDropDevice(board, drop_pins)
+        pd_control = PurpleDropController(pd_dev, board, ecal)
     else:
         print("Launching HW server...")
         pd_dev = PersistentPurpleDropDevice()
