+++ setPwm.py
@@ -14,28 +14,28 @@ def main(chan, duty_cycle):
     msg = SetPwmMsg()
     msg.chan = chan
     msg.duty_cycle = duty_cycle
-    
+
     devices = list_purpledrop_devices()
     if(len(devices) == 0):
         print("No PurpleDrop USB device found")
         sys.exit(1)
-    elif len(devices) > 1: 
+    elif len(devices) > 1:
         print("Multiple PurpleDrop devices found. Please ammend software to allow selection by serial number")
         for d in devices:
             print(f"{d.device}: Serial {d.serial_number}")
         sys.exit(1)
     print(f"Connecting to purple drop on {devices[0].device}")
     dev = PurpleDropDevice(devices[0].device)
-    listener = dev.get_sync_listener(msg_filter=CommandAckMsg)
-    print("Sending Message\n")
-    dev.send_message(msg)
-    print("Waiting for ack\n")
-    ack = listener.wait(timeout=1.0)
+    with dev.get_sync_listener(msg_filter=CommandAckMsg) as listener:
+        print("Sending Message\n")
+        dev.send_message(msg)
+        print("Waiting for ack\n")
+        ack = listener.next(timeout=1.0)
     if ack is None:
         print("No ACK message received")
     else:
         print("Got ACK: " + str(ack))
 
-    
+
 if __name__ == '__main__':
     main()
